Fix external links in documentation.
We had a bunch of dead links in the documentation. I went through and fixed up a bunch of them.
I didn't address all of them, as some resources just can't be found anymore, but I didn't want to erase evidence of their existence.

The other thing I didn't really touch is a number of wiki.rosettacommons.org links to resources which possibly could be (have already been)
moved over to the Gollum documentation wiki.